+++ src/5_OsExp/src/gdt/gdt.asm
@@ -1,5 +1,3 @@
-; Don't use
-
 ; Sources:
 ; https://www.cs.bham.ac.uk/~exr/lectures/opsys/10_11/lectures/os-dev.pdf
 ; page 34
