+++ src/5_OsExp/stdlib/stdport.c
@@ -1,0 +1,5 @@
+#include "stdport.h"
+
+void outPortb(unsigned short _port, unsigned char _data) {
+    __asm__ __volatile__ ("outb %1, %0" : :  "dN" (_port), "a" (_data));
+}
