+++ src/5_OsExp/stdlib/stdport.h
@@ -1,0 +1,2 @@
+// Writes value to specified hardware port.
+void outPortb(unsigned short _port, unsigned char _data);
