+++ src/5_OsExp/stdlib/stdtxt.h
@@ -1,0 +1,2 @@
+// Function to determine the length of a buffer.
+int strlen(unsigned char *string);
